requirements.txt
moviepy==2.2.1
yt-dlp==2025.5.22
flask==3.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
flask-cors==4.0.0
pytz==2023.3